% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.wts.R
\name{barplot.wts}
\alias{barplot.wts}
\title{barplot.wts}
\usage{
\method{barplot}{wts}(
  wts.efront,
  legend.text = NULL,
  col = NULL,
  ylab = NULL,
  xlab = c("MU", "VOL"),
  bar.ylim = NULL,
  ...
)
}
\arguments{
\item{wts.efront}{}

\item{legend.text}{}

\item{col}{}

\item{ylab}{}

\item{xlab}{}

\item{bar.ylim}{}

\item{...}{}
}
\value{

}
\description{
barplot.wts
}
