% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotWts.R
\name{barplotWts}
\alias{barplotWts}
\title{barplotWts}
\usage{
barplotWts(
  wts.efront,
  legend.text = NULL,
  col = NULL,
  ylab = NULL,
  xlab = c("MU", "VOL"),
  bar.ylim = NULL
)
}
\arguments{
\item{wts.efront}{}

\item{legend.text}{}

\item{col}{}

\item{ylab}{}

\item{xlab}{}

\item{bar.ylim}{}
}
\value{

}
\description{
barplotWts
}
