% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2Functions.R
\name{barplotWtsDoug}
\alias{barplotWtsDoug}
\title{Title}
\usage{
barplotWtsDoug(
  wts.efront,
  legend.text = NULL,
  col = NULL,
  ylab = NULL,
  xlab = "VOL",
  bar.ylim = NULL,
  ...
)
}
\arguments{
\item{wts.efront}{}

\item{legend.text}{}

\item{col}{}

\item{ylab}{}

\item{xlab}{}

\item{bar.ylim}{}

\item{...}{}
}
\value{

}
\description{
Title
}
