% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEfronts.R
\name{bootEfronts}
\alias{bootEfronts}
\title{bootEfronts}
\usage{
bootEfronts(returns, pspec, rf = 0.03, npoints = 20, B = 3,
  Seed = NULL, gmv = T, maxSR = F, xlim = NULL, ylim = NULL,
  k.sigma = 2, k.mu = 2, digits = 4, figTitle = NULL)
}
\arguments{
\item{returns}{}

\item{pspec}{}

\item{rf}{}

\item{npoints}{}

\item{B}{}

\item{Seed}{}

\item{gmv}{}

\item{maxSR}{}

\item{xlim}{}

\item{ylim}{}

\item{k.sigma}{}

\item{k.mu}{}

\item{digits}{}

\item{figTitle}{}
}
\description{
bootEfronts
}
