% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.Efront.R
\name{chart.Efront}
\alias{chart.Efront}
\title{chart.Efront}
\usage{
chart.Efront(
  returns,
  pspec,
  firstEfront = T,
  gmv = T,
  maxSR = T,
  rf = 0,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  n.portfolios = 10
)
}
\arguments{
\item{returns}{}

\item{pspec}{}

\item{firstEfront}{}

\item{gmv}{}

\item{maxSR}{}

\item{rf}{}

\item{xlim}{}

\item{ylim}{}

\item{xlab}{}

\item{ylab}{}

\item{n.portfolios}{}
}
\value{

}
\description{
chart.Efront
}
