% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIV.R
\name{DIV}
\alias{DIV}
\title{DIV is the function to calculate the diversification index for a set of portfolio weights. DIV is one minus the HHI index.}
\usage{
DIV(weights)
}
\arguments{
\item{weights}{}
}
\value{

}
\description{
DIV is the function to calculate the diversification index for a set of portfolio weights. DIV is one minus the HHI index.
}
