% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divHHI.R
\name{divHHI}
\alias{divHHI}
\title{divHHI is the function to calculate the diversification index for a set of portfolio weights. DIV is one minus the HHI index.}
\usage{
divHHI(weights)
}
\arguments{
\item{weights}{}
}
\value{

}
\description{
divHHI is the function to calculate the diversification index for a set of portfolio weights. DIV is one minus the HHI index.
}
