% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfPointMassFig.R
\name{edfPointMassFig}
\alias{edfPointMassFig}
\title{This is code for plotting a point mass distribution}
\usage{
edfPointMassFig(r, x1 = -0.02, x2 = 0.02)
}
\arguments{
\item{r}{A vector of values for the plot x axis}

\item{x1}{Lower limite of plotting range}

\item{x2}{Upper limit of plotting range}
}
\value{

}
\description{
This is code for plotting a point mass distribution
}
\examples{
edfPointMassFig(.005)
}
