% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efront2Asset.R
\name{efront2Asset}
\alias{efront2Asset}
\title{Efficient Frontier for 2-Asset Portfolio}
\usage{
efront2Asset(wts, rho, muVol = c(0.2, 0.1, 0.15, 0.04))
}
\arguments{
\item{wts}{}

\item{rho}{}

\item{muVol}{}
}
\value{

}
\description{
Efficient Frontier for 2-Asset Portfolio
}
