% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efrontplot.shiny.R
\name{efrontplot.shiny}
\alias{efrontplot.shiny}
\title{efrontplot.shiny}
\usage{
efrontplot.shiny(files, group, list.cset, list.arg)
}
\arguments{
\item{files}{}

\item{group}{}

\item{list.cset}{}

\item{list.arg}{}
}
\value{

}
\description{
efrontplot.shiny
}
