% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMetrics.R
\name{etl}
\alias{etl}
\title{NORMALIZED (DEFAULT) AND NON-NORMALIZED ETL}
\usage{
etl(R, alpha = 0.05, normalize = TRUE)
}
\arguments{
\item{R}{}

\item{alpha}{}

\item{normalize}{}
}
\value{

}
\description{
NORMALIZED (DEFAULT) AND NON-NORMALIZED ETL
}
