% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewmaMeanVol.R
\name{ewmaMeanVol}
\alias{ewmaMeanVol}
\title{Exponential Weighted Moving Average (EWMA) Mean Volitility}
\usage{
ewmaMeanVol(
  x,
  nstart = 10,
  robMean = T,
  robVol = T,
  cc = 2.5,
  lambdaMean = 0.9,
  lambdaVol = 0.9,
  Dyn = F,
  lambdaMeanDyn = 0.7,
  lambdaVolDyn = 0.7
)
}
\arguments{
\item{x}{returns of the portfolio}

\item{nstart}{}

\item{robMean}{if the robust mean is used, default is T}

\item{robVol}{if the robust vol is used, default is T}

\item{cc}{}

\item{lambdaMean}{}

\item{lambdaVol}{}

\item{Dyn}{}

\item{lambdaMeanDyn}{}

\item{lambdaVolDyn}{}
}
\value{

}
\description{
Exponential Weighted Moving Average (EWMA) Mean Volitility
}
\details{
The robust EWMA mean algorithm has the form
}
