% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewmaVol.R
\name{ewmaVol}
\alias{ewmaVol}
\title{Exponential Weighted Moving Average (EWMA) Volitility}
\usage{
ewmaVol(x, nstart = 10, robVol = T, cc = 2.5, lambda = 0.9)
}
\arguments{
\item{x}{returns of the portfolio}

\item{nstart}{}

\item{robVol}{if the robust vol is used, default is T}

\item{cc}{}

\item{lambda}{}
}
\value{

}
\description{
Exponential Weighted Moving Average (EWMA) Volitility
}
