% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTsfmStats.R
\name{fitTsfmStats}
\alias{fitTsfmStats}
\title{fitTsfmStats}
\usage{
fitTsfmStats(x, rsq = F, digits = 3, show.signif.stars = TRUE)
}
\arguments{
\item{x}{}

\item{rsq}{}

\item{digits}{}

\item{show.signif.stars}{}
}
\value{

}
\description{
fitTsfmStats
}
