% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFitStats.R
\name{lmFitStats}
\alias{lmFitStats}
\title{lmFitStats}
\usage{
lmFitStats(x, rsq = F, digits = 3, show.signif.stars = TRUE)
}
\arguments{
\item{x}{}

\item{rsq}{}

\item{digits}{}

\item{show.signif.stars}{}
}
\description{
lmFitStats
}
