% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathEfront.R
\name{mathEfront}
\alias{mathEfront}
\title{mathEfront}
\usage{
mathEfront(returns, mu.max = NULL, sigma.max = NULL, rf = 0.005,
  rf.line = T, stocks = T, stock.names = T, values = T,
  identify.stock.names = F, npoints = 100, digits = NULL)
}
\arguments{
\item{returns}{}

\item{mu.max}{}

\item{sigma.max}{}

\item{rf}{}

\item{rf.line}{}

\item{stocks}{}

\item{stock.names}{}

\item{values}{}

\item{identify.stock.names}{}

\item{npoints}{}

\item{digits}{}
}
\description{
mathEfront
}
