% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathEfrontCashRisky.R
\name{mathEfrontCashRisky}
\alias{mathEfrontCashRisky}
\title{mathEfrontCashRisky}
\usage{
mathEfrontCashRisky(
  returns,
  rf = 0.005,
  scalex = 1.2,
  scaley = 2,
  display.stocks = T,
  stock.names = T,
  risk.tol = F,
  npoints = 10,
  plot.efront = T,
  wts.plot = T,
  equal.wts = T,
  bar.ylim = c(0, 1),
  digits = NULL
)
}
\arguments{
\item{returns}{}

\item{rf}{}

\item{scalex}{}

\item{scaley}{}

\item{display.stocks}{}

\item{stock.names}{}

\item{risk.tol}{}

\item{npoints}{}

\item{plot.efront}{}

\item{wts.plot}{}

\item{equal.wts}{}

\item{bar.ylim}{}

\item{digits}{}
}
\value{

}
\description{
mathEfrontCashRisky
}
