% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathEfrontRisky.R
\name{mathEfrontRisky}
\alias{mathEfrontRisky}
\title{mathEfrontRisky}
\usage{
mathEfrontRisky(
  returns,
  npoints = 100,
  efront.only = T,
  display = T,
  digits = NULL
)
}
\arguments{
\item{returns}{}

\item{npoints}{}

\item{efront.only}{}

\item{display}{}

\item{digits}{}
}
\value{

}
\description{
mathEfrontRisky
}
