% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathTport.R
\name{mathTport}
\alias{mathTport}
\title{mathTport}
\usage{
mathTport(returns, rf = 0.01, digits = NULL)
}
\arguments{
\item{returns}{}

\item{rf}{}

\item{digits}{}
}
\description{
mathTport
}
