% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathWtsEfrontRisky.R
\name{mathWtsEfrontRisky}
\alias{mathWtsEfrontRisky}
\title{mathWtsEfrontRisky}
\usage{
mathWtsEfrontRisky(
  returns,
  mu.efront,
  bar.ylim = c(0, 1),
  display.wts = T,
  digits = NULL
)
}
\arguments{
\item{returns}{}

\item{mu.efront}{}

\item{bar.ylim}{}

\item{display.wts}{}

\item{digits}{}
}
\value{

}
\description{
mathWtsEfrontRisky
}
