% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2Functions.R
\name{mathWtsEfrontRiskyMuCov}
\alias{mathWtsEfrontRiskyMuCov}
\title{Title}
\usage{
mathWtsEfrontRiskyMuCov(muRet, volRet, corrRet, mu.efront, digits = NULL)
}
\arguments{
\item{muRet}{}

\item{volRet}{}

\item{corrRet}{}

\item{mu.efront}{}

\item{digits}{}
}
\value{

}
\description{
Title
}
