% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathWtsEfrontRiskyNew.R
\name{mathWtsEfrontRiskyNew}
\alias{mathWtsEfrontRiskyNew}
\title{mathWtsEfrontRiskyNew}
\usage{
mathWtsEfrontRiskyNew(returns, mu.efront, display.wts = T, digits = NULL)
}
\arguments{
\item{returns}{}

\item{mu.efront}{}

\item{display.wts}{}

\item{digits}{}
}
\value{

}
\description{
mathWtsEfrontRiskyNew
}
