% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumActMgr.R
\name{muEst}
\alias{muEst}
\title{muEst returns a simple weighted average of the current return and the last period's return.}
\usage{
muEst(r, mu0, sigma0, win_level = 4, lambda = 0.9)
}
\arguments{
\item{r}{Current period return}

\item{mu0}{Mean return from the last period}

\item{sigma0}{Estimate of volatility from the last period. If unavailable, set sigma0 = 0}

\item{win_level}{Number of standard deviations at which we winsorize (default: win_level =4)}

\item{lambda}{Exponential weighting constant (default: lambda = 0.9)}
}
\description{
muEst returns a simple weighted average of the current return and the last period's return.
}
