% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.outputMeanVol.R
\name{opt.outputMeanVol}
\alias{opt.outputMeanVol}
\title{opt.outputMeanVol}
\usage{
opt.outputMeanVol(opt, returns, digits = NULL, names = NULL)
}
\arguments{
\item{opt}{}

\item{returns}{}

\item{digits}{}

\item{names}{}
}
\value{

}
\description{
opt.outputMeanVol
}
