% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.outputMvo.R
\name{opt.outputMvo}
\alias{opt.outputMvo}
\title{opt.outputMvo}
\usage{
opt.outputMvo(opt, returns, digits = NULL, names = NULL, rf = 0)
}
\arguments{
\item{opt}{}

\item{returns}{}

\item{digits}{}

\item{names}{}

\item{rf}{}
}
\value{

}
\description{
opt.outputMvo
}
