% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberRhoPsiWtsEff.R
\name{psiHuber}
\alias{psiHuber}
\title{Huber Psi}
\usage{
psiHuber(x, cc = 1.345)
}
\description{
Huber Psi
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{huber}
