% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumActMgr.R
\name{sigmaEst}
\alias{sigmaEst}
\title{sigmaEst computes a EW estimate of volatility}
\usage{
sigmaEst(r, mu0, sigma0, win_level = 4, lambda_in = 0.9,
  lambda_out = 0.8)
}
\arguments{
\item{r}{Current period return}

\item{mu0}{Prior period mean return (often set to 0 in finance as market efficiency diminishes the mean)}

\item{sigma0}{Prior period estimated volatility. If unavailable (e.g. in the first period), sigma0 = 0}

\item{win_level}{Number of standard deviations at which we winsorize (default: 4)}

\item{lambda_in}{EW constant when the data seems consistent with the current estimate of volatility (default: 0.9)}

\item{lambda_out}{EW constant when the data seems inconsistent with the current estimate of volatility (default: 0.8)}
}
\description{
sigmaEst computes a EW estimate of volatility
}
