% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMetrics.R
\name{starrRatio}
\alias{starrRatio}
\title{NORMALIZED (DEFAULT) AND NON-NORMALIZED STARR}
\usage{
starrRatio(R, alpha = 0.05, normalize = TRUE, Rf = 0)
}
\arguments{
\item{R}{}

\item{alpha}{}

\item{normalize}{}

\item{Rf}{}
}
\value{

}
\description{
NORMALIZED (DEFAULT) AND NON-NORMALIZED STARR
}
