% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnoverOpt.R
\name{turnoverOpt}
\alias{turnoverOpt}
\title{turnoverOpt}
\usage{
turnoverOpt(
  returns,
  mu.target = NULL,
  wts.initial,
  toc,
  long.only = TRUE,
  printD = F,
  printA = F,
  printout = T
)
}
\arguments{
\item{returns}{}

\item{mu.target}{}

\item{wts.initial}{}

\item{toc}{}

\item{long.only}{}

\item{printD}{}

\item{printA}{}

\item{printout}{}
}
\value{

}
\description{
turnoverOpt
}
