% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnoverSimple.R
\name{turnoverSimple}
\alias{turnoverSimple}
\title{turnOverSimple}
\usage{
turnoverSimple(wts, wtsInit = NULL, digits = NULL)
}
\arguments{
\item{wts}{}

\item{wtsInit}{}

\item{digits}{}
}
\value{

}
\description{
turnOverSimple
}
