% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberRhoPsiWtsEff.R
\name{wtsHuber}
\alias{wtsHuber}
\title{Huber Wts}
\usage{
wtsHuber(x, cc = 1.345)
}
\description{
Huber Wts
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{huber}
