% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2Functions.R, R/edfPointMassFig.R
\name{edfPointMassFig}
\alias{edfPointMassFig}
\title{Title}
\usage{
edfPointMassFig(r, x1 = -0.02, x2 = 0.02)

edfPointMassFig(r, x1 = -0.02, x2 = 0.02)
}
\arguments{
\item{r}{}

\item{x1}{}

\item{x2}{}

\item{r}{A vector of values for the plot x axis}

\item{x1}{Lower limite of plotting range}

\item{x2}{Upper limit of plotting range}
}
\description{
Title

This is code for plotting a point mass distribution
}
\examples{
edfPointMassFig(.005)
}
