% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2Functions.R, R/mathEfrontRisky.R
\name{mathEfrontRiskyMuCov}
\alias{mathEfrontRiskyMuCov}
\title{Title}
\usage{
mathEfrontRiskyMuCov(muRet, volRet, corrRet, npoints = 100,
  display = T, efront.only = T, digits = NULL)

mathEfrontRiskyMuCov(muRet, volRet, corrRet, npoints = 100,
  display = T, efront.only = T, digits = NULL)
}
\arguments{
\item{muRet}{}

\item{volRet}{}

\item{corrRet}{}

\item{npoints}{}

\item{display}{}

\item{efront.only}{}

\item{digits}{}

\item{print}{}

\item{muRet}{}

\item{volRet}{}

\item{corrRet}{}

\item{npoints}{}

\item{display}{}

\item{efront.only}{}

\item{digits}{}
}
\description{
Title

mathEfrontRiskyMuCov
}
