% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2Functions.R
\name{mathWtsEfrontRisky}
\alias{mathWtsEfrontRisky}
\title{Title}
\usage{
mathWtsEfrontRisky(returns, mu.efront, digits = NULL)
}
\arguments{
\item{returns}{}

\item{mu.efront}{}

\item{digits}{}
}
\description{
Title
}
