% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathWtsEfrontRisky.R
\name{mathWtsEfrontRiskyMuCov}
\alias{mathWtsEfrontRiskyMuCov}
\title{mathWtsEfrontRiskyMuCov}
\usage{
mathWtsEfrontRiskyMuCov(muRet, volRet, corrRet, mu.efront,
  efront.only = T, display.wts = T, npoints = 5, digits = NULL)
}
\arguments{
\item{muRet}{}

\item{volRet}{}

\item{corrRet}{}

\item{mu.efront}{}

\item{efront.only}{}

\item{display.wts}{}

\item{npoints}{}

\item{digits}{}
}
\description{
mathWtsEfrontRiskyMuCov
}
